price_estimator/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vinyl-price-estimator"
version = "0.1.0"
description = "Vinyl price estimation models"
requires-python = ">=3.12"
dependencies = [
    "vinyl-shared",
    "botasaurus>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
    "requests>=2.28.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "mlflow>=2.9.0",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["price_estimator*"]

[tool.uv.sources]
vinyl-shared = { workspace = true }
